refactor(*): router、config、雪花算法的重构
- router重构了method掩码的定义格式
- config单独封装一个包并增加Redis配置项
- snowflake封装到generator包，做了一些命名的修改，允许用户手动创建多个配置不同的生成器
