+++ gs/id_generator.go
@@ -1,0 +1,7 @@
+package gs
+
+import (
+	"github.com/dan-kuroto/gin-stronger/generator"
+)
+
+var SnowFlake = generator.NewSnowFlake(Config)
